backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
python-docx==1.1.2
pdfplumber==0.11.4
sentence-transformers==3.1.1
scikit-learn==1.5.2
numpy==2.1.2